calculator 고정 소수점 연산 적용완료 (주석 제거x 및 target_stock_price 함수 테스트 필요

+++ src/types.rs
@@ -9,7 +9,7 @@ pub enum Leverage {
 }
 
 impl Leverage {
-    pub fn value(&self) -> i32 {
+    pub fn value(&self) -> i64 {
         match self {
             Leverage::Daily2x => 2,
             Leverage::Daily3x => 3,
